# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dispatch API
  version: 3.0.0
  description: |
    Open-source bulk email campaign platform with tracking, multi-provider support, and batch processing.

    > **Note:** This spec is maintained by hand and may lag the implementation. Treat the live API
    > (and the route handlers under `apps/api/src/routes`) as the source of truth. Browse it interactively
    > at `/docs` (Swagger UI) on any running instance.

    ## Authentication
    Most endpoints require authentication via a `session_token` cookie obtained from the login endpoint.
    Alternatively, an API key can be passed via the `X-API-Key` header.

    ## CSRF Protection
    All mutation endpoints (POST, PUT, DELETE) require a valid `X-CSRF-Token` header.
    The CSRF token is issued as a cookie on every response.

    ## Common Response Format
    All endpoints return a standard envelope:
    ```json
    {
      "success": true,
      "data": { ... },
      "message": "Human-readable message",
      "meta": { "pagination": { ... } }
    }
    ```
  contact:
    name: Dispatch
    url: https://github.com/iamjpsingh/bulk-email-sender
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Session token cookie set on login/register.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token required for POST/PUT/DELETE requests. Value is read from the `csrf_token` cookie.

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (shape varies per endpoint).
        message:
          type: string
        meta:
          type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
      required: [success]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Something went wrong
        error:
          type: string
      required: [success, message]
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasMore:
          type: boolean
          example: true

    # -- Auth --
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
      required: [id, email, name]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
          format: password
          minLength: 8
      required: [email, name, password]

    # -- Config --
...
security:
  - cookieAuth: []
  - apiKeyAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Account created. Session cookie set.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
        '400':
          description: Validation error or email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful. Session cookie set.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /send:
    post:
      tags: [Send]
      summary: Send emails to a list of contacts
      operationId: sendEmails
      description: |
        Accepts a multipart form with an Excel file of contacts, HTML content, and sending options.
        Supports immediate, batched, and scheduled sending modes.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                excelFile:
                  type: string
                  format: binary
                  description: Excel file (.xlsx) with contact data.
                subject:
                  type: string
                htmlContent:
                  type: string
                  description: HTML body content (from WYSIWYG editor).
                htmlTemplate:
                  type: string
                  format: binary
                  description: Optional HTML template file (overrides htmlContent).
                configId:
                  type: string
                  description: SMTP config ID to use. Falls back to default.
                delay:
                  type: integer
                  default: 20
                  description: Delay in seconds between emails (non-batch mode).
                useBatch:
                  type: string
                  enum: ['on']
                  description: Enable batch mode.
                batchSize:
                  type: integer
                  default: 20
                batchDelay:
                  type: integer
                  default: 60
                  description: Delay in seconds between batches.
                emailDelay:
                  type: integer
                  default: 45
                  description: Delay in seconds between emails within a batch.
                scheduleEmail:
                  type: string
                  enum: ['on']
                  description: Enable scheduling.
                scheduledTime:
                  type: string
                  format: date-time
                  description: ISO 8601 datetime for scheduled send.
                notifyEmail:
                  type: string
                  format: email
                  description: Email address for job completion notification.
                emailRangeStart:
                  type: integer
                  default: 0
                  description: Starting index in the contact list.
                emailRangeCount:
                  type: integer
                  description: Number of contacts to send to.
              required: [excelFile, subject]
      responses:
        '200':
          description: Sending initiated or job scheduled.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          jobId:
                            type: string
                          contactCount:
                            type: integer
                          configUsed:
                            type: string
                          batchMode:
                            type: boolean
                          scheduledMode:
                            type: boolean
                          scheduledTime:
                            type: string
                            format: date-time
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /parse-excel:
    post:
      tags: [Send]
      summary: Parse an Excel file and preview contacts
      operationId: parseExcel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                excelFile:
                  type: string
                  format: binary
              required: [excelFile]
      responses:
        '200':
          description: Parsed contacts preview (first 5 rows) and total count.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          contacts:
                            type: array
                            items:
                              type: object
                            maxItems: 5
                          totalCount:
                            type: integer
...